fix: refactor(core-repo): harden reference-data path resolution and tighten UI import boundary
